+++ flutter_feature_biometric_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+* add parameter `encodedKey` inside `isBiometricChanged`
+* change `key` to `alias` inside `authenticateSecureEncrypt`
+* change `key` to `alias` inside `authenticateSecureDecrypt`
+* change `encodedIVKey` to `encodedKey` inside `authenticateSecureDecrypt`
+* change `AuthenticatorStatus` to `FeatureAuthenticatorStatus`
+
 ## 0.1.0
 
 * Update Docs